local: Remove custom code to set scale in the data format structure
Remove the code reading the scale into the format structure. This is now
done better by iio_context_update_scale_offset(), which is called at the
end of iio_create_context(), and has the added advantage of reading the
scale for all channels (not just the scan-elements) and also read the
offsets.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>