context: Update channels' scale/offset after creating context
The iio_data_format.scale field was previously only set for scan-element
channels. This was wrong, as regular channels can have a scale as well.
Both can also have an offset, which is added to the raw value before
applying the scale; this offset was not present in the iio_data_format
structure.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>